рефакторит шаблон редактирования и добаления точки маршрута объединяя их в один
